Move code to build attribute(s) to a separate class.
